requirements.txt
mindsdb[postgresql]
pytest
psycopg2-binary
